sightseeingMap内にSelectedTravelSpotを呼び出す
